+++ wechat_django/admin/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from . import views, wechatapp
+from . import views, wechatapp # noqa