+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 
 ## 0.3.0
 * 微信支付client
-* 统一下单订单管理及信号通知
+* 统一下单订单管理,回调及订单状态变更信号
+* 小程序的client由`wechat_django.client.WeChatClient`变更为`wechatpy.client.api.WeChatWxa`
+* 数个配置项更改
+* 站点代码重构
 
 ## 0.2.5
 * 模板消息